requirements(1).txt
transformers
torch
pillow
gradio